replacing `float` with `decimal.Decimal` to avoid rounding errors

+++ packages/valory/skills/price_estimation_abci/behaviours.py
@@ -19,6 +19,7 @@
 
 """This module contains the behaviours for the 'abci' skill."""
 
+from decimal import Decimal
 from abc import ABC
 from typing import Generator, Optional, Set, Type, cast
 
@@ -89,7 +90,7 @@ def to_int(most_voted_estimate: float, decimals: int) -> int:
     decimal_places = most_voted_estimate_[::-1].find(".")
     if decimal_places > decimals:
         most_voted_estimate_ = most_voted_estimate_[: -(decimal_places - decimals)]
-    most_voted_estimate = float(most_voted_estimate_)
+    most_voted_estimate = Decimal(most_voted_estimate_)
     int_value = int(most_voted_estimate * (10 ** decimals))
     return int_value
